RHOAIENG-65096: Add connection type extension points (Phase 2)
Define extension point types in plugin-core for connection type
service (data fetching) and UI component (form fields, details helper,
default value renderer, form) extensions. Move useWatchConnectionTypes
and useServingConnections hooks to ui-core backed by resolved
extensions. Internalize model-serving-specific components
(ConnectionOciAlert, ConnectionOciPathField, ConnectionS3FolderPathField,
LabeledConnection) into the model-serving package. Register host-side
extension providers in frontend/src/plugins/extensions/connection-types.ts.
Create re-export shims for backward compatibility — consumers continue
importing through @odh-dashboard/internal; migration to direct extension
consumption will follow separately.
