+++ frontend/src/pages/modelServing/screens/projects/InferenceServiceModal/ConnectionOciAlert.tsx
@@ -1,13 +1,2 @@
-import { Alert } from '@patternfly/react-core';
-import React from 'react';
-
-export const ConnectionOciAlert: () => JSX.Element = () => (
-  <Alert
-    variant="warning"
-    style={{ marginTop: 'var(--pf-t--global--spacer--lg)' }}
-    title="Model requirement for OCI connections"
-  >
-    OCI connections can be used to deploy only models packaged as ModelCars. Deploying models
-    packaged in different formats will result in failure.
-  </Alert>
-);
+// eslint-disable-next-line @odh-dashboard/no-restricted-imports -- re-exporting from model-serving for backward compatibility
+export { ConnectionOciAlert } from '@odh-dashboard/model-serving/components/connectionTypes/ConnectionOciAlert';
