+++ frontend/src/pages/modelServing/screens/types.ts
@@ -7,7 +7,6 @@ import type {
   ServingContainer,
   ServingRuntimeKind,
 } from '@odh-dashboard/model-serving/shared';
-import { Connection } from '#~/concepts/connectionTypes/types';
 import { EnvVariableDataEntry } from '#~/pages/projects/types';
 
 export enum PerformanceMetricType {
@@ -89,7 +88,5 @@ export type ServingPlatformStatuses = {
   refreshNIMAvailability: () => Promise<boolean | undefined>;
 };
 
-export type LabeledConnection = {
-  connection: Connection;
-  isRecommended?: boolean;
-};
+// eslint-disable-next-line @odh-dashboard/no-restricted-imports -- re-exporting from model-serving for backward compatibility
+export type { LabeledConnection } from '@odh-dashboard/model-serving/components/connectionTypes/types';
