+++ frontend/src/pages/projects/screens/detail/connections/useServingConnections.ts
@@ -1,29 +1,2 @@
-import React from 'react';
-import { FetchState } from '@odh-dashboard/ui-core/hooks/useFetchState';
-import { Connection } from '#~/concepts/connectionTypes/types';
-import { isModelServingCompatible } from '#~/concepts/connectionTypes/utils';
-import useConnections from './useConnections';
-
-const useServingConnections = (
-  namespace?: string,
-  includeDashboardFalse = false,
-  skipCompatibilityCheck = false,
-): FetchState<Connection[]> => {
-  const {
-    data: connections,
-    loaded,
-    error,
-    refresh,
-  } = useConnections(namespace, undefined, includeDashboardFalse);
-  return React.useMemo(
-    () => [
-      connections.filter((c) => skipCompatibilityCheck || isModelServingCompatible(c)),
-      loaded,
-      error,
-      refresh,
-    ],
-    [connections, loaded, error, refresh, skipCompatibilityCheck],
-  );
-};
-
-export default useServingConnections;
+// eslint-disable-next-line @odh-dashboard/no-restricted-imports -- re-exporting from ui-core for backward compatibility
+export { default } from '@odh-dashboard/ui-core/hooks/useServingConnections';
