+++ frontend/src/plugins/extensions/index.ts
@@ -1,8 +1,14 @@
 import type { Extension } from '@openshift/dynamic-plugin-sdk';
+import connectionTypeExtensions from './connection-types';
 import navigationExtensions from './navigation';
 import routeExtensions from './routes';
 import taskExtensions from './tasks';
 
-const extensions: Extension[] = [...navigationExtensions, ...routeExtensions, ...taskExtensions];
+const extensions: Extension[] = [
+  ...connectionTypeExtensions,
+  ...navigationExtensions,
+  ...routeExtensions,
+  ...taskExtensions,
+];
 
 export default extensions;
