+++ packages/k8s-core/src/index.ts
@@ -136,6 +136,7 @@ export type {
   Connection,
   ConnectionTypeFormData,
   AWSDataEntry,
+  FieldMode,
 } from './connectionTypes';
 
 export { NamespaceApplicationCase, byName, namespaceToProjectDisplayName } from './projectUtils';
