+++ packages/model-serving/src/components/connectionTypes/types.ts
@@ -1,0 +1,6 @@
+import type { Connection } from '@odh-dashboard/k8s-core';
+
+export type LabeledConnection = {
+  connection: Connection;
+  isRecommended?: boolean;
+};
