+++ packages/ui-core/package.json
@@ -15,6 +15,8 @@
     "./hooks/useBrowserStorage": "./src/hooks/useBrowserStorage.tsx",
     "./hooks/useFetch": "./src/hooks/useFetch.ts",
     "./hooks/useFetchState": "./src/hooks/useFetchState.ts",
+    "./hooks/useWatchConnectionTypes": "./src/hooks/useWatchConnectionTypes.ts",
+    "./hooks/useServingConnections": "./src/hooks/useServingConnections.ts",
     "./design": "./src/design/index.ts",
     "./design/vars.scss": "./src/design/vars.scss",
     "./design/icons": "./src/images/icons/index.ts",
@@ -67,6 +69,7 @@
   "dependencies": {
     "@odh-dashboard/foundation": "*",
     "@odh-dashboard/k8s-core": "*",
+    "@odh-dashboard/plugin-core": "*",
     "classnames": "^2.5.1",
     "lodash-es": "^4.18.1"
   },
@@ -78,6 +81,7 @@
     "@types/node": "^22.0.0"
   },
   "peerDependencies": {
+    "@openshift/dynamic-plugin-sdk": "^5",
     "@patternfly/react-core": "^6",
     "@patternfly/react-icons": "^6",
     "@patternfly/react-styles": "^6",
